iox-#1391 Move 'static_storage' to public include and adapt namespace
