+++ iceoryx_hoofs/include/iceoryx_hoofs/cxx/function.hpp
@@ -17,8 +17,8 @@
 #ifndef IOX_HOOFS_FUNCTION_HPP
 #define IOX_HOOFS_FUNCTION_HPP
 
-#include "iceoryx_hoofs/internal/cxx/static_storage.hpp"
 #include "iceoryx_hoofs/internal/cxx/storable_function.hpp"
+#include "iceoryx_hoofs/memory/static_storage.hpp"
 
 namespace iox
 {
